refactor: Consolidate workout feedback context and hooks
- Refactored WorkoutFeedbackContext to streamline the structure and improve maintainability by removing redundant code and custom hooks.
- Updated imports in WorkoutDetailPage to utilize the new hooks directory for accessing workout feedback, enhancing clarity and organization.
- Improved ErrorBoundary component by updating environment variable checks for better compatibility with the current setup.
