+++ src/hooks/index.ts
@@ -26,3 +26,4 @@ export {
 export * from "./useWorkoutInstances";
 export { usePusherBeamsInterests } from "./usePusherBeamsInterests";
 export { usePusherEvent } from "./usePusherEvent";
+export * from "./useWorkoutFeedback";
